pyproject.toml
[tool.poetry]
name = "indoor-api"
version = "0.1.0"
description = ""
authors = ["Marcos Pulgar <marcos.pulgar.a@gmail.com>"]
readme = "README.md"
packages = [{include = "indoor_api"}]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "0.95.0"
mysql-connector-python = "8.0.33"
peewee = "3.16.2"
python-dotenv = "1.0.0"
pymysql = "^1.0.3"
mangum = "^0.17.0"
pytz = "^2023.3.post1"
decouple = "^0.0.7"


[tool.poetry.group.dev.dependencies]
pytest-asyncio = "^0.21.0"
pytest-mock = "^3.10.0"
coverage = "^7.2.7"
devtools = "^0.11.0"
uvicorn = "^0.22.0"
httpx = "^0.24.1"
pytest-env = "^0.8.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
